Move sketcher utils functions to new App/Utils.
