+++ src/Mod/Sketcher/App/SketchObject.cpp
@@ -80,8 +80,8 @@
 #include <Mod/Part/App/DatumFeature.h>
 #include <Mod/Part/App/GeometryMigrationExtension.h>
 #include <Mod/Part/App/TopoShapeOpCode.h>
-#include <Mod/Sketcher/Gui/Utils.h>
 
+#include "Utils.h"
 #include "SketchObject.h"
 #include "SketchObjectPy.h"
 #include "SolverGeometryExtension.h"
